Rename the modules in the llm_chat subpackage
