+++ mongo_analyser/llm_chat/openai.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Iterator, List, Optional
 
 import openai
 
-from .base_chat import LLMChat
+from .base import LLMChat
 
 logger = logging.getLogger(__name__)
